+++ PluginsAndFeatures/azure-toolkit-for-rider/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - Do not update App Services during the deployment ([RIDER-119626](https://youtrack.jetbrains.com/issue/RIDER-119626))
+- Search for the `in-proc` folders for the in-process worker model ([RIDER-120960](https://youtrack.jetbrains.com/issue/RIDER-120960))
 
 ## [4.3.7] - 2024-12-09